Added day model and adapter
Each day is a cell in the grid that fills the calendar, for each booking on that day a line is displayed
