+++ app/src/main/AndroidManifest.xml
@@ -2,7 +2,7 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="com.hva.m2mobi.m2hva_reservationsystem">
 
-    <uses-permission android:name="android.permission.INTERNET"></uses-permission>
+    <uses-permission android:name="android.permission.INTERNET" />
     <uses-permission android:name="android.permission.WRITE_CALENDAR" />
     <uses-permission android:name="android.permission.READ_CALENDAR" />
     <uses-permission android:name="android.permission.GET_ACCOUNTS" />
@@ -25,19 +25,19 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-            android:label="@string/title_activity_tab" />
+        android:label="@string/title_activity_tab" />
         <activity
             android:name=".activities.LoginActivity"
             android:label="@string/title_activity_login" />
-        <activity android:name=".activities.RoomsOverviewActivity"></activity>
-        <activity android:name=".activities.RoomDetailActivity">
-        </activity>
-       <activity android:name=".activities.ReserveRoomActivity">
-       </activity>
+        <activity android:name=".activities.RoomsOverviewActivity" />
+        <activity android:name=".activities.RoomDetailActivity"></activity>
+        <activity android:name=".activities.ReserveRoomActivity"></activity>
         <activity
             android:name=".activities.TabActivity"
-            android:label="@string/title_activity_tab"></activity>
-        <activity android:name=".activities.AccountAuthorisationActivity"></activity>
+            android:label="@string/title_activity_tab" />
+        <activity android:name=".activities.AccountAuthorisationActivity" />
+        <activity android:name=".activities.CalendarTestActivity" />
         <activity android:name=".activities.CalendarActivity"></activity>
     </application>
+
 </manifest>
